Moved and cleanup.
